pyproject.toml
[project]
name = "my-board-job"
version = "0.1.0"
description = ""
authors = [
    {name = "Anh Dung Nguyen",email = "nad1186@gmail.com"}
]
package-mode = false
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.2,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "djangorestframework-api-key (>=3.1.0,<4.0.0)",
    "pytest (>=8.4.0,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ipdb = "^0.13.13"
